backend/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
eventlet>=0.35.0
numpy>=1.24.0
pydantic>=2.0.0
cryptography>=41.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
